refactor app from gui into new controler
